Add files via upload
This file allows for over the web (OTA) updates. By first uploading the OTALastMinuteEngineersWeb file onto the board using a cable, it is able to intercept OTA updates. It also provides you with an IP address. Then, you can run the blink sketch example to have the esp32 create a website. on this website accessed with the ip address, you can then upload the .bin file to change the code remotely.